Run tests only if tests specified.
If you have 2 projects and only one has defined tests then TestRunner is
going to atempt to run [project_name]_tests.fif even if non generated.
Firt exception is going to be thrown.
So i think the best solution is just to not enter test processing loop
if there are no tests.